requirements.txt
suds
dnspython
